requirements/requirements.txt
Faker
pandas
google-cloud-storage
apache-airflow